Update all the things! Camera smooth-moves.
The font is changed, just as a test, to a non-square one, Iosevka Slab;
if you can widen the window, it will become more square. The camera
glides smoothly through a 300x90 map, following the player, and
displaying 100x30 tiles at a time. The player also moves smoothly, using
SquidLayers.slide(). This does mean that the player renders over the
contents of a cell, which may need some minor work to correct. The stats
window is currently not displayed; it won't be too hard to get it to
appear again.

Some build files changed while I was trying to get this to build; the
problem was on my side and relates to Java 9 prereleases being hot
garbage.