Bring back smooth-move because colors affect it
Colors have also changed to a less-contrasting scheme for walls and
floors. The background is SColor.DB_INK instead of true black, which has
a faint blue tint but is otherwise just different enough from black to
avoid the flickering effect.

Also, pathfinding improvements based on stuff that's already the default
in SquidLib's demos, using a precalculated DijkstraMap that can be
stepped through until the move finishes (when monsters are added, they
can interrupt movement when first seen maybe?).