Performance improvements on larger map
There seems to be no jitter now on movement even when the map is
300x100, thanks to a series of not-especially-obvious optimizations. We
reuse the existing fovResult array instead of creating a new one each
time we move. We update existing RememberedTile objects instead of
creating many new ones (it discarded old ones, stressing GC) each time
we move. In a key and subtle trick, we refill the blocked GreasedRegion
with the small visible area and then get its "fringe" (the area just
outside the visible area, using 8-way adjacency) to form a wall of
path-blocking cells. This last one avoids filling a GreasedRegion with
many thousands of blocked Coords, which DijkstraMap would potentially
have to iterate through on each scan. Also, the Dijkstra scan is limited
to 1.45 times the SIGHT stat of the player, where 1.45 gives enough
leeway for the cases where circular FOV is larger than the amount
Dijkstra can normally reach (this step may not be needed, or a smaller
number might be helpful).