Get info to use smaller, sharper font, fix resize()
I think resize() was broken before this commit, but I'm not sure. The
font is the same face, but a different size and has different smoothing.
This uses way more pixels in the info bar and figures out what to put
for a pixel using a 4x4 pixel area for each pixel used; this is handled
relatively cleanly by StretchViewport.