Minor changes for big speedup. Camera fix.
The animation duration is increased in this commit because everything is
now being processed faster during render(), so we don't need to use such
a low delay and visually benefit from a larger delay (animations
actually appear; if the animation duration kept at 0.13, then they would
go by too quickly). The camera has been fixed a bit; this seems to be
dependent on the size of the dungeon map (not its visual bounds), and
now it avoids a jerking when approaching the edges of the scrolled area.