Clear up when we need to putMap()
The comments were misleading before, sorry about that. Because the
SquidLayers still knows what colors and chars it has, you only need to
putMap() when those change, and only need to updateStats() when the
stats change. The mouse movement needs to update the map, but not stats,
while actions (including any kinds of input) update both.

This also changes the head of the text-art armor person to a capital
Omega, to look more helm-like.