Make font in context/info smaller, simplify layout
The font for context and info is the same but now the shader is
different in SquidLib, so we don't really benefit from drawing to a
larger area and shrinking it down with viewports. We do, however, get to
use smaller font sizes. This means the context and info viewports don't
cover more pixels in a smaller area anymore, and they are now
normal-sized. The TextCellFactory is different for the smaller font and
the smoothing multiplier needed a lot of fine-tuning.

Also, I fixed a typo in the word Berserk (etymology is the same as a
very old word for bear, probably "ber", and "sark", a pre-Old-English
term for a shirt, due to berserkers wearing bear hides and fighting with
the ferocity of bears in battle).