Bring back creatures, avoiding removal-while-using
The earlier code (before the SparseLayers rewrite) created an Actor each
time a creature moved, and destroyed it when it was no longer moving.
This causes issues for some reason with SparseLayers, where if a Glyph
is destroyed while the list  of Glyphs is being iterated over, the
iteration goes out of bounds. To save on the amount of garbage we
create, a Glyph belongs to each Physical that we try to move (it is null
if it never moves), and we make it visible while moving it or set its
appearance to be a ' ' char if we aren't using it currently. It isn't
the fastest, but it looks more than smooth enough.