Messages do the scroll thing
The message area is larger now, and displays 4 rows of messages. Some
messages were printed many times when the mouse moved; these have been
taken out to avoid floods of messages. Old messages are not currently
stored, which allows the messages to be stored in a small array instead
of an arbitrarily-large list.

LightRNG usage has switched to ThrustRNG; both allow skip() to be called
with the same arguments to do the same thing, but Thrust should be
somewhat faster and should have comparable-or-better quality.