Reorganize project so IDEA stops complaining
I was getting a persistent issue every time  I opened the project
because the assets folder was shared between multiple modules for some
reason. It's only needed in desktop currently, so that's where they have
been moved. This also fixes a bug where a monster could spawn in the
player's tile, and makes monsters not block FOV (which exposed some
counter-intuitive quirks of shadowcasting FOV). The desktop project is
back to using LWJGL2 because I think Java 9 has issues with LWJGL3,
though I don't even have Java 9 installed to check. One of the RNGs now
uses a ThrustAltRNG instead of a ThrustRNG because ThrustAlt has very
high quality and isn't much different on speed (between the faster
ThrustRNG and the slower LightRNG).