Fix inventory bug where drawing didn't remove...
Before, drawing a weapon wouldn't remove it from inventory, but it would
add your previously-drawn weapon back to the inventory. This resulted in
an extra item in the inventory for every time the 'd' key was pressed.
Now, items are removed from inventory when they are moved to being
equipped.

SquidLib is updated.

Weapons have also had some element improvements for the pre-Modification
versions.