Various adjustments to lighting speed, noisiness
The calculations involving the lights variable only need to be done
once, so they've been moved to create() instead of called every time we
putMap(). The lighting is more chaotic in how it grows and shrinks now,
in part due to using a higher multiplier on time when calculating noise,
and in part due to a tricky adjustment in putWithLight that subtracts a
smoothly-changing, but quickly-changing, number based on noise from the
normal lightAmount calculation.