Allow seeding all generation with Epigon.rootChaos
Normally the seed only applies to some things, like the map and enemy
placement, but for weapon types/elements to be determined as well, you
can change Epigon's rootChaos field to use a seed, which will make all
weapons appear the same every time that seed is used.

This commit also includes some work that may be used soon to reduce the
amount of StatefulRNGs we have by just storing their states, which are
very lightweight.