Very basic scrollbar shows position next to arrows
This is only in the help screen for now. It should be a little easier to
bring this into other screens than it was to write it this time.
Utilities' capitalization stuff changed too; there were some issues that
wouldn't have showed up in the limited usage of caps(), but would in any
general usage (trailing spaces would be stripped by String.split(), no
way to control the delimiter used in the input vs. in the output, there
was a bug where the delimiter given did nothing, etc.).