Fixed some minimap placement/updating bugs
Remembered tiles now appear on the minimap again, and the minimap is
placed correctly. There's a possible SquidLib bug in mouse handling that
was worked around in a hacky way before, and that hack ceased to work
when grid sizes changed. In resize(), for contextMouse's
reinitialization, `-(int) (infoSLayers.getTop() + 8f)` only works when
the cell height, adjusted for zoom, is 8 pixels. Now it uses the height
of a cell to adjust, and infoMouse does this too, which fixes the bug at
least somewhat. A bug was also fixed in the minimap where an array could
have its bounds exceeded when rendering a Glyph that was removed from
mapSLayers' glyphs list. Another bug related to double-updating Glyph
creatures with act() in the minimap, it's also fixed.