Conditions show as an overlay character on a Glyph
When a character is suffering from some status condition (which still
does nothing right now), a glyph appears over them (a `~` for the player
and a `˝` for enemies) in an appropriate color for the type of attack
that inflicted it. I'm thinking about how to rework some stuff so combat
modifications are separated out somewhat, reversible (for status effects
when they end), and compatible with existing stuff.