Make weapons one of many possible statEffects
So, statEffects is now a field on Physical, and it stores things like
non-innate bonuses from equipment or similarly-temporary penalties from
ailments. All statEffects are tallied up in ActionOutcome and applied to
a _copy_ of the creature's actual stats, both for actor and target,
before running to-hit/damage-dealt calculations. These changes shouldn't
persist between actions.