Fix wielding (weapons weren't applied before)
Weapons had been made so they had a ChangeTable for the stat changes
they should cause when wielded, but that ChangeTable wasn't ever being
used because OrderedSet.alter() doesn't do anything when the key isn't
present (I forgot this). Now we add the ChangeTable if not present, or
alter it if it is; this may be equivalent to OrderedSet.add() if order
doesn't matter, but it can in the presence of limits on LiveValues.
Also, because LiveValues are mutable, they need to be deep-copied when
they are values in a data structure; without a deep copy, I had a
playthrough where the player dealt about 10 more damage every time he
attacked, until each musket shot dealt 90+ damage... At least now we
know this handles large values for stats.