Allow wielding multiple weapons, flailing wildly
Each weapon has a likelihood it will be used based on how many hands it
takes to wield, and your unarmed fighting style or element of magic
counts as an infrequent option even when wielding a weapon or weapons.
You can hold two one-handed weapons and have a 40% chance of attacking
with the left hand, 40% with the right hand, or a 20% chance of
attacking unarmed; you can hold a two-handed weapon and have an 80%
chance of attacking with it and a 20% of attacking unarmed.

ChangeTable now also can calculate multiple changes to LiveValues
without hitting the min/max clamps until it applies the final value.