Various presentation improvements for conditions
They can now choose which overlay character to place over the attached
creature, if any, as well as the specific element used for a Condition
that may differ from its blueprint. The element determines the color of
the overlay, as well. The groundwork is set to start adding Conditions!
I still need to make constructing Effects easier, since those are used
during ticks.