As an alternative to Gradle, mavenize (Java 9 only)
The idea is to potentially use Moditect (
https://github.com/moditect/moditect ), which is at this point in time
Maven-only, to add Java 9 module descriptors to SquidLib (which is a
Java 7 project) and libGDX (which is Java 6) so they can maybe be used
to form a jlink modular runtime image, with a much smaller distributable
size and faster startup than a full bundled JRE. Not sure how possible
this is. The current version doesn't get quite as good FPS as before,
though still over 200 FPS. I've tried it with OpenJ9 (former IBM JRE)
and with OpenJDK (mostly the same as Oracle), and OpenJDK seems a fair
bit better, for unclear reasons. OpenJ9 may be better on memory usage,
but I'm not sure.