Font looks better, but reduces FPS somewhat
I get about 120-140 FPS now with the MSDF font instead of 160-180 with
the SDF font, but chars like '(' render without chunks getting taken out
of them (an issue with the SDF) and the shapes look pretty sharp. It's
up to @SquidPony what hardware should be targeted; very very old
processors with no GPU (like a 10-or-more-year-old Core 2 Duo) may be
unable to render the larger MSDF texture since it's 4096x4096. We don't
target Android, so the major issue of Android devices that only support
2048x2048 or less doesn't affect us. Whatever fonts we don't use should
be moved out of assets/ or deleted; they are bundled in distributed jars
regardless of whether they are used.