Add Redux project (next to Proto) to test removal
Proto has been updated to the latest SquidLib, and then copied to a
parallel project in the same repo, Redux. Several old classes have been
removed from Redux (such as those for dungeon gen, which were mostly
empty and all unused, as well as DataMaster, which was virtually unused
and can be added back if needed). The idea is I can freely remove
classes from Redux, but if something goes wrong, I still have Proto to
get them back in the project. EpiData is slightly different now,
calculating a unique int for the first 2 to the 32 objects created and
only producing an id String from that int if requested (it then caches
it). This could be reworked as a long and its uniqueness integrated with
the chaos variables used by other classes.