Make 0-handed but equipped weapons use a slot
There's a wreath that uses the head slot, a cape and a relic that use
the neck slot, and new bladed boots that use the feet slot. These items
can now be sundered, and you can't equip multiple items in the same slot
(no boots over boots). The clothing slot types may need to be unified
with the wield slot types.

The seed is fixed currently so bladed boots spawn nearby; they look like
a backwards 'L'.