Work on some weapons, preparing for ranged combat
Since the contents of Iosevka-Family-msdf are slightly different from
Iosevka (non-family) as we had at the start, the full contents of the
current font are in Redux/Iosevka-Family-contents-msdf.txt , and may
need to be checked against the char listing in Utilities (probably in
some automated way).

PositionRNG is no longer a RandomnessSource (it was never used by RNG),
just in case there was some polymorphic issue slowing down RNGs with
different RandomnessSources.