Try out different determine() in GauntRNG
Not sure how this will work, relative to LightRNG.determine(), but it
could be a little faster and the quality seems excellent so far (no
anomalies on 1TB, but still running). Mostly I want to have this code
preserved in a commit in case I roll back to LightRNG but want the code
for future development.

Epigon.rootChaos is back to being unseeded.