Fix a bug where weapons stayed after sundering
Before, if your weapon was destroyed it would (sometimes?) stay in the
table of possible things you could attack with, leading to weird cases
where your unarmed fighter was doing stabbing or vacuous damage without
using magic (he had used a vacuous bow before, but it was sundered).
This seems fixed, and the change also applies to disarming as well as
sundering.