Scroll messages, more GauntRNG experiments
The big thing here is being able to scroll to earlier messages with
PageUp and scroll back down with PageDown (it automatically jumps to the
bottom if a new message comes in). This also updates the SquidLib dep,
which may help pathfinding performance a good amount (it could be
insignificant).