Start refactor from GauntRNG to EpiData as an RNG
This should use the same amount of memory, since the state is still one
long (still called chaos), it'll just be in EpiData instead of Physical.
Each Physical acted as an RNG anyway before, it just needed GauntRNG to
know what to do with its state. Now we can use existing SquidLib methods
that use IRNG instances instead of jury-rigging GauntRNG to fill in
their places.