Switch PositionRNG to use Mizuchi variant
Mizuchi is just LinnormRNG with a stream it adds instead of always adding 1 in its first step; it's in Sarong but not SquidLib yet. This selects the stream using a slight variant on Linnorm/Mizuchi to randomize the influence of x and y jointly when the PositionRNG moves. Notably, this doesn't set the state to a number based on stream anymore if you call move(long, long); it will always use a state of 0, which seems good in testing. A few of the tests show a decent improvement over the old way, but most are similar in quality (the new way should be faster since Mizuchi is almost exactly as fast as Linnorm, which is our fastest good generator, in Sarong's benchmarks).