Stairs and multiple levels!
Up is '≤', like a rewind icon, down is '≥' like a play or fast forward icon. Press '<' on a '≤' tile, or press '>' on a '≥' tile, to go on the stairs. It takes you to the same coordinates in the dungeon, but a different floor, with a staircase back at your position.