Slots are almost unified
Clothing, Jewelry, and OverArmor slots all now implement BodySlot, which extends ConstantKey. This just means that you can have all equipment in one Map with BodySlot keys, and BodySlot adds a location on the text-art "stick figure" for where to show damage to or emanation from an equipped item.

This updates TabLab to fix a bug and starts to add creature data, though it isn't used yet.