Fix staircases, path/menu overlap
Staircases were broken because the check for if you were on a staircaase was first checking the blocking contents of the currrent tile, and that now includes the player if he's there. The path to cursor line was showing up even when a menu was open to choose a weapon; now it doesn't appear. The lightEmitted fields on Physical and Modification are no longer needed because we have Radiance, so they were removed. Physical's equals() and hashCode() methods now compare most of Physical's fields and two of EpiData's fields, and are probably much slower but will allow Physical.unique to be checked with a simple UnorderedSet, probably. EpiData now uses a different, sorta-experimental RNG that I'm still testing.