Falling mode works with 'r' on map
Also, the win screen for falling mode shows the count of gold coins you gathered and also the counts of any gems you gathered twice or more.

This also fixes a crash bug in falling mode from the last commit, and re-enables being able to enter it (it was '<' before, now it's 'r').