FOV has several possible shapes
These use a new feature of FOV in SquidLib that seems to work fine here. The LOS technique used before doesn't block enough behind a player if there are lights in that area, so now LOS is limited to twice sight distance, which is then multiplied by the fractional directional range (which is sometimes 0.0 for behind). Try hitting ctrl-shift-backspace a few times to reload the player and see different FOV shapes.