Switch EpiData random to Tangle
This is an ideal usage for TangleRNG, since having many objects with different starting states means they'll all have different sets of possible numbers they can and can't generate.

Also, this commit makes it possible to enter values for "trainings" in RawCreature.tsv that don't correspond to any enum values in Skill; if none match it will act like it was "brawler". The numbers still need to be 0-7 because they correspond to Rating values.