Remove directional FOV. Add beam weapons.
The directional FOV was very disorienting and caused play with the keyboard to act differently than play with the mouse (keys could move into a direction you couldn't see, but the mouse shouldn't potentially reveal a path where you can't see). Also, there's counterfeit lightsabers in various shapes using a beam weapon modification. The player always starts with a beam weapon, and it's identical to a normal weapon except it strobes with a really bright glow.