Adjust EpiData, remove posrng
The code for PositionRNG is still there, but the static one in Epigon.java is no longer needed. Because there's a Physical in every tile so far (the floor), we can use it as a random number generator for that cell, which is only logical. PositionRNG was only used to generate numbers in EpiTile, which has floors anyway,