Check in latest small changes before attempting refactor
I'm going to try to take out subcell stuff because I think it slows down rendering a lot when there are many (sub)cells to update. With about 1/15 of the cells emitting light,  I get at most 21 FPS after some warmup when not much is running in the background (I am using 50% of my CPU on an RNG test almost all the time, so keep that in mind). 21 FPS is kinda acceptable, but it doesn't control smoothly at that rate, and slide effects are noticeably jerky and awkward.

@squidpony , if you instead want to keep subcell rendering, please send some kind of message today (October 30) before I start ripping stuff apart.