Walnuts and pecans will freeze or ignite you, affecting color warmth
Trees are 3x more often here; change WorldGenerator's "9 in 256" line to get it back. This uses the YCwCm color space (which I came up with and so far really like) to tweak colors towards cold colors and a frosted look when chilled and warm colors (with flickering) when on fire. Walnuts are technically Frost Walnuts and pecans are Roaster's Pecans; they each permanently raise a stat slightly when used but injure you by the condition they inflict.