Made a quick test to show a libGDX issue; will move to a better place soon.
GridPoint2 in libGDX has a rather bad hashCode() implementation that will collide more than 50% of the time when enough GridPoint2 objects are used in an ObjectMap or ObjectSet; "enough objects" can be as low as 150x150, and 200x200 will run out of heap usually. This is partly an ObjectMap bug but mostly a GridPoint2 bug, since giving it just a tiny bit better of a hashCode() fixes this. Thanks to @darencius for putting up with my questions about this bug; he also noted that using java.util.HashMap will do fine with GridPoint2 as-is.