Partial fix to determinism breakage
This commit makes the game startup fully deterministic in every way, where before the level layout was deterministic but the contents and player creation were random. Before, we used Epigon.rng (which has a fixed seed) to set up levels, and used either the random number generation in each EpiData or Epigon.rootChaos (which had a random seed before, and a fixed seed now) to give the player his equipment, place items and enemies, determine enemy dialogue, etc. It would be good if we can figure out where we're introducing random aspects to the game.