GDX ObjectMap's cuckoo hashing is just worse than linear probing.
I copied the oldest implementation of ObjectMap from libGDX, before it used cuckoo hashing, and that ancient version (present as OldMap) doesn't run out of heap on medium-small keysets of a type with a bad hashCode(). It seems to take a bit longer than SquidLib's UnorderedMap, but the code is much smaller. I did need to make two changes to OldMap to get it to correctly resize and for the index masking to work.