Some optimization for rain.
Now instead of the color varying slightly over each brief frame (which I couldn't actually notice in play), each raindrop and associated puddle gets a slightly different color. You can fix the element of the rain to be always water on line 2366 of Epigon.java . The main reason this hits framerate so hard is probably just the 20K-50K Actions it was adding before; now it's down to 3K-8K and the duration is shorter. That should be good, because input to the game is locked during the animation.