OPTIONAL: Raise JDK to 16
If this causes trouble, you can just revert it. If you want to use the new Java 16 features (desktop-only, like Epigon), you should make sure you have Java 16 installed and that Gradle is currently using Java 16 for itself (this can be because Java 16 is your JAVA_HOME installation, or because you configured IDEA to make Gradle use the project JDK for Gradle).